move helper methoe to session helper
