+++ app/controllers/application_controller.rb
@@ -1,4 +1,5 @@
 class ApplicationController < ActionController::API
+  include SessionsHelper
   before_action :authorized
   def encode_token(payload)
     JWT.encode(payload, 's3cr3t')
@@ -17,20 +18,5 @@ class ApplicationController < ActionController::API
         nil
       end
     end
-  end
-
-  def current_user
-    if decoded_token
-      user_id = decoded_token[0]['user_id']
-      @user = User.find_by(id: user_id)
-    end
-  end
-
-  def logged_in?
-    !!current_user
-  end
-
-  def authorized
-    render json: { message: 'Please log in to peform this operation' }, status: :unauthorized unless logged_in?
   end
 end
