+++ app/helpers/sessions_helper.rb
@@ -1,0 +1,16 @@
+module SessionsHelper
+  def current_user
+    if decoded_token
+      user_id = decoded_token[0]['user_id']
+      @user = User.find_by(id: user_id)
+    end
+  end
+
+  def logged_in?
+    !!current_user
+  end
+
+  def authorized
+    render json: { message: 'Please log in to peform this operation' }, status: :unauthorized unless logged_in?
+  end
+end
